+++ tests/FSharp.Data.Tests/XmlProvider.fs
@@ -1210,17 +1210,27 @@ let ``simple types are formatted properly``() =
     isValid minValues |> should equal true
     isValid maxValues |> should equal true
 
+
 [<Test>]
-let ``time is invalid for xs:date``() =
-    let simpleValues =
+let ``time is omitted when zero``() =
+    let simpleValues date = 
       SimpleTypes.A(
         int = 0,
         long = 0L,
-        date = System.DateTime.Today.AddHours(1.),
+        date = date,
         dateTime = System.DateTimeOffset.Now,
         boolean = false,
         decimal = 0M,
         double = System.Double.NaN)
         .ToString()
     let isValid = isValid SimpleTypesXsd
-    isValid simpleValues |> should equal false
+
+    let validXml = System.DateTime(2018, 8, 29) |> simpleValues
+    isValid validXml |> should equal true
+    (XElement.Parse validXml).Attribute(XName.Get "date").Value
+    |> should equal "2018-08-29"
+
+    let invalidXml = System.DateTime(2018, 8, 29, 5, 30, 56) |> simpleValues
+    isValid invalidXml |> should equal false
+    (XElement.Parse invalidXml).Attribute(XName.Get "date").Value
+    |> should equal "2018-08-29T05:30:56.0000000"